icerberux/src/main/webapp/app/entities/cerbac-when/delete/cerbac-when-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CerbacWhenService } from '../service/cerbac-when.service';

import { CerbacWhenDeleteDialogComponent } from './cerbac-when-delete-dialog.component';

describe('Component Tests', () => {
  describe('CerbacWhen Management Delete Component', () => {
    let comp: CerbacWhenDeleteDialogComponent;
    let fixture: ComponentFixture<CerbacWhenDeleteDialogComponent>;
    let service: CerbacWhenService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [CerbacWhenDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(CerbacWhenDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(CerbacWhenDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(CerbacWhenService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
